chg: adaptações para as mudanças de identificadores das tabelas

+++ src/main/java/com/project/sp_medical_group/Jpa/Repositories/EnderecoJpaRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnderecoJpaRepository extends JpaRepository<Endereco, Integer> { }
+public interface EnderecoJpaRepository extends JpaRepository<Endereco, Long> { }
