+++ src/main/java/com/project/sp_medical_group/Jpa/Repositories/MedicoJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicoJpaRepository extends JpaRepository<Medico, Integer> {
+public interface MedicoJpaRepository extends JpaRepository<Medico, Long> {
 }
