+++ src/main/java/com/project/sp_medical_group/Models/Especialidade.java
@@ -21,7 +21,7 @@ public class Especialidade {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Setter(AccessLevel.NONE)
     @Column(name = "especialidade_id")
-    private Integer especialidadeId;
+    private Long especialidadeId;
 
     private String nome;
